% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMDdesign.R
\name{BMDdesign}
\alias{BMDdesign}
\title{Find optimal designs for benchmark dose analysis}
\usage{
BMDdesign(
  grad_fun,
  dr_fun,
  cvec = NULL,
  obj,
  theta,
  d0 = NULL,
  n0 = NULL,
  N1 = NULL,
  num_doses,
  max_dose,
  swarm,
  iter,
  alg,
  ignore_stage1 = T,
  show_progress = T
)
}
\arguments{
\item{grad_fun}{Dose-response gradient function.}

\item{dr_fun}{Dose-response function. Used to account for non-constant variance.}

\item{cvec}{c vector for finding c-optimal designs.}

\item{obj}{Design objective. Can either be "c" or "D".}

\item{theta}{Vector of local model parameter values.}

\item{d0}{Doses in prior experiment.}

\item{n0}{Sample size allocations in prior experiment.}

\item{N1}{Sample size allowance for current experiment. Used when finding augmented designs.}

\item{num_doses}{Number of doses in the design.}

\item{max_dose}{Maximum allowed dose.}

\item{swarm}{Swarm size for swarm-based optimization algorithm.}

\item{iter}{Maximum number of iterations for optimization algorithm.}

\item{alg}{Algorithm from metaheuristicOpt}

\item{ignore_stage1}{If TRUE, finds the locally optimal design.}

\item{show_progress}{If TRUE, show progress bar for optimization.}
}
\description{
Find optimal designs for benchmark dose analysis
}
\examples{
# find designs for logistic model ###########################################
# parameter estimates obtained using ToxicR on deguelin data from drc package
theta_logistic = c(-1.5627368, 0.1258373)

# D-optimal design
set.seed(1136)
logistic_Dopt = BMDdesign(
  grad_fun = logistic.grad,
  dr_fun = logistic.fun,
  obj = 'D',
  theta = theta_logistic,
  num_doses = 2,
  max_dose = 50.11872,
  swarm = 20,
  iter = 100,
  alg = 'DE'
)
logistic_Dopt$x
logistic_Dopt$w
check_eq(logistic_Dopt)

# c-optimal design for BMD at BMR=0.1
cvec = logistic.bmdgrad(0.1, theta_logistic)
set.seed(1140)
logistic_copt = BMDdesign(
  grad_fun = logistic.grad,
  dr_fun = logistic.fun,
  obj = 'c',
  cvec = cvec,
  theta = theta_logistic,
  num_doses = 2,
  max_dose = 50.11872,
  swarm = 30,
  iter = 500,
  alg = 'DE'
)
logistic_copt$x
logistic_copt$w
check_eq(logistic_copt)

# 3-parameter Weibull model #################################################
# parameter estimates obtained using ToxicR on deguelin data from drc package
theta_weibull = c(-1.145036439, 1.804710767, 0.004269448) # note g is reparameterized

# D-optimal design
set.seed(1152)
weibull_Dopt = BMDdesign(
  grad_fun = weibull.grad,
  dr_fun = weibull.fun,
  obj = 'D',
  theta = theta_weibull,
  num_doses = 3,
  max_dose = 50.11872,
  swarm = 20,
  iter = 500,
  alg = 'DE'
)
weibull_Dopt$x
weibull_Dopt$w
check_eq(weibull_Dopt)

# c-optimal design
cvec = weibull.bmdgrad(0.1, theta_weibull)
set.seed(1156)
weibull_copt = BMDdesign(
  grad_fun = weibull.grad,
  dr_fun = weibull.fun,
  obj = 'c',
  cvec = cvec,
  theta = theta_weibull,
  num_doses = 3,
  max_dose = 50.11872,
  swarm = 20,
  iter = 500,
  alg = 'DE'
)
weibull_copt$x
weibull_copt$w
check_eq(weibull_copt)
}
